producer/pyproject.toml
[tool.poetry]
name = "producer"
version = "0.1.0"
description = ""
authors = ["InnoCentCoder5497  <innocentcoder5497@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
my-script = "producer.producer_module:main"
start-producer = "producer.producer_module:stream_plant_data"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python = "^2.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
